Split up headers and implementation
so we eventually get a compilation firewall
