+++ test/libintervalxt/induction.benchmark.cc
@@ -21,7 +21,7 @@
 #include <benchmark/benchmark.h>
 #include <gtest/gtest.h>
 
-#include <intervalxt/iet.hpp>
+#include <intervalxt/interval_exchange_transformation.hpp>
 
 using namespace intervalxt;
